requirements.txt
streamlit>=1.35.0

supabase>=2.4.0
gotrue>=2.4.0

qrcode[pil]>=7.4.2
streamlit-qrcode-scanner>=0.0.8

pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2

plotly>=5.20.0

scikit-learn>=1.4.0

prophet>=1.1.5

reportlab>=4.1.0
Pillow>=10.3.0

cryptography>=42.0.0

python-dotenv>=1.0.1
loguru>=0.7.2
httpx>=0.27.0
